Solution for Task 1 from Practice 9
In this suggested solution we keep the capacity (maximum elements that array can store). If push back a new element when the capacity is already full, we double the current capacity in order to save some expensive operations like "resizing" array for adding next elements. We have a "test" with a single push back and pop back of element. It is supposed that the program works correctly when we push back/pop back consistently more than one element